Made a pretty bad equivariant constraint
Besides that, I made the shape of x_train and x_test (28, 28, 1) from
the beginning, instead of having to add it a bunch of random places.

I also changed the network structure a good bit. This might not be the
final iteration.

The way the equivariant constraint works is not ideal. For instance, it
will not allow the identity matrix.