Equivariant CNN, small polishes, speed improvements (#7)
* Made a pretty bad equivariant constraint

Besides that, I made the shape of x_train and x_test (28, 28, 1) from
the beginning, instead of having to add it a bunch of random places.

I also changed the network structure a good bit. This might not be the
final iteration.

The way the equivariant constraint works is not ideal. For instance, it
will not allow the identity matrix.

* Delete some old rotation code

Now we only use the rotations from tfa, avoiding pillows and the CPU

* Minor changes, changed network

Changed the network so the architecture fits better with the plan. Also,
add an option for mstep and such

* Improvements to network, more tf.function

* Last bit of work

* made pylint happy

* Fix tests

* Various fixes

Co-authored-by: Mantzius <magnni18@student.aau.dk>